ADEN-11569 mapper => extender
